requirements.txt
spacy>=3.5
pillow
pytesseract
PyMuPDF==1.22.3
tqdm
python-dotenv
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
streamlit